Changed how value/type symbol info is stored, and basic hover support
At the moment, the distinction between names that refer to values vs names that refer to types is inconsistent. Before this commit, whether a name was referred to as a type or a value was decided by the context it appears in, which any visitor of a tree knows on its own. However, things like the LSP's hover handler don't have this context - so providing this information inside the Tree.Name itself is needed. But now the symbol's context has two different ways of getting it.